make environmental hazards use floats for strength to allow less than 1 ppm per operation to be output

+++ src/main/java/com/gregtechceu/gtceu/api/machine/feature/IEnvironmentalHazardEmitter.java
@@ -27,7 +27,7 @@ public interface IEnvironmentalHazardEmitter extends IMachineFeature {
     /**
      * @return the starting strength of the hazard zone.
      */
-    int getHazardStrengthPerOperation();
+    float getHazardStrengthPerOperation();
 
     default void spreadEnvironmentalHazard() {
         if (!ConfigHolder.INSTANCE.gameplay.environmentalHazards) {
